My commit

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,9 +7,9 @@
  */
 void print_number(int n)
 {
-	unsigned int m; /* power of 10 */
+	long m; /* power of 10 */
 	int c; /* boolean check */
-	unsigned int num; /* convert int to long */
+	long num; /* convert int to long */
 
 	num = n;
 	/* negatives */
